+++ pulsar/pulsar_test.go
@@ -29,7 +29,7 @@ import (
 	ecdsahelper "github.com/insolar/insolar/cryptohelpers/ecdsa"
 	"github.com/insolar/insolar/log"
 	"github.com/insolar/insolar/pulsar/entropygenerator"
-	"github.com/insolar/insolar/pulsar/pulsartestutil"
+	"github.com/insolar/insolar/pulsar/pulsartestutils"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -61,12 +61,12 @@ func TestNewPulsar_WithoutNeighbours(t *testing.T) {
 	mockListener := func(connectionType string, address string) (net.Listener, error) {
 		actualConnectionType = connectionType
 		actualAddress = address
-		return &pulsartestutil.MockListener{}, nil
+		return &pulsartestutils.MockListener{}, nil
 	}
-	storage := &pulsartestutil.MockPulsarStorage{}
+	storage := &pulsartestutils.MockPulsarStorage{}
 	storage.On("GetLastPulse", mock.Anything).Return(&core.Pulse{PulseNumber: 123}, nil)
 	clientFactory := &MockRpcClientFactoryWrapper{}
-	clientFactory.On("CreateWrapper").Return(&pulsartestutil.MockRPCClientWrapper{})
+	clientFactory.On("CreateWrapper").Return(&pulsartestutils.MockRPCClientWrapper{})
 
 	result, err := NewPulsar(configuration.Configuration{
 		Pulsar: configuration.Pulsar{
@@ -77,7 +77,7 @@ func TestNewPulsar_WithoutNeighbours(t *testing.T) {
 	},
 		storage,
 		clientFactory,
-		pulsartestutil.MockEntropyGenerator{},
+		pulsartestutils.MockEntropyGenerator{},
 		nil,
 		mockListener,
 	)
@@ -86,7 +86,7 @@ func TestNewPulsar_WithoutNeighbours(t *testing.T) {
 	assert.Equal(t, privateKey, result.PrivateKey)
 	assert.Equal(t, "testType", actualConnectionType)
 	assert.Equal(t, "listedAddress", actualAddress)
-	assert.IsType(t, result.Sock, &pulsartestutil.MockListener{})
+	assert.IsType(t, result.Sock, &pulsartestutils.MockListener{})
 	assert.NotNil(t, result.PrivateKey)
 
 	clientFactory.AssertNumberOfCalls(t, "CreateWrapper", 0)
@@ -103,10 +103,10 @@ func TestNewPulsar_WithNeighbours(t *testing.T) {
 
 	expectedPrivateKey, _ := ecdsahelper.GeneratePrivateKey()
 	parsedExpectedPrivateKey, _ := ecdsahelper.ExportPrivateKey(expectedPrivateKey)
-	storage := &pulsartestutil.MockPulsarStorage{}
+	storage := &pulsartestutils.MockPulsarStorage{}
 	storage.On("GetLastPulse", mock.Anything).Return(&core.Pulse{PulseNumber: 123}, nil)
 	clientFactory := &MockRpcClientFactoryWrapper{}
-	clientFactory.On("CreateWrapper").Return(&pulsartestutil.MockRPCClientWrapper{})
+	clientFactory.On("CreateWrapper").Return(&pulsartestutils.MockRPCClientWrapper{})
 
 	result, err := NewPulsar(
 		configuration.Configuration{
@@ -122,8 +122,8 @@ func TestNewPulsar_WithNeighbours(t *testing.T) {
 		},
 		storage,
 		clientFactory,
-		pulsartestutil.MockEntropyGenerator{}, nil, func(connectionType string, address string) (net.Listener, error) {
-			return &pulsartestutil.MockListener{}, nil
+		pulsartestutils.MockEntropyGenerator{}, nil, func(connectionType string, address string) (net.Listener, error) {
+			return &pulsartestutils.MockListener{}, nil
 		})
 
 	assertObj.NoError(err)
@@ -136,7 +136,7 @@ func TestNewPulsar_WithNeighbours(t *testing.T) {
 func TestPulsar_EstablishConnection_IsInitialised(t *testing.T) {
 	pulsar := &Pulsar{Neighbours: map[string]*Neighbour{}}
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("IsInitialised").Return(true)
 	mockClientWrapper.On("Lock")
 	mockClientWrapper.On("Unlock")
@@ -158,7 +158,7 @@ func TestPulsar_EstablishConnection_IsInitialised(t *testing.T) {
 func TestPulsar_EstablishConnection_IsNotInitialised_ProblemsCreateConnection(t *testing.T) {
 	pulsar := &Pulsar{Neighbours: map[string]*Neighbour{}}
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("IsInitialised").Return(false)
 	mockClientWrapper.On("CreateConnection", mock.Anything, mock.Anything).Return(errors.New("test reasons"))
 	mockClientWrapper.On("Lock")
@@ -182,10 +182,10 @@ func TestPulsar_EstablishConnection_IsNotInitialised_ProblemsWithRequest(t *test
 	pulsar := &Pulsar{
 		Neighbours:       map[string]*Neighbour{},
 		PrivateKey:       mainPrivateKey,
-		EntropyGenerator: pulsartestutil.MockEntropyGenerator{},
+		EntropyGenerator: pulsartestutils.MockEntropyGenerator{},
 	}
 
-	mockClientWrapper := &pulsartestutil.CustomRPCWrapperMock{}
+	mockClientWrapper := &pulsartestutils.CustomRPCWrapperMock{}
 
 	done := make(chan *rpc.Call, 1)
 	done <- &rpc.Call{Error: errors.New("oops, request is broken")}
@@ -208,10 +208,10 @@ func TestPulsar_EstablishConnection_IsNotInitialised_SignatureFailed(t *testing.
 	pulsar := &Pulsar{
 		Neighbours:       map[string]*Neighbour{},
 		PrivateKey:       mainPrivateKey,
-		EntropyGenerator: pulsartestutil.MockEntropyGenerator{},
+		EntropyGenerator: pulsartestutils.MockEntropyGenerator{},
 	}
 
-	mockClientWrapper := &pulsartestutil.CustomRPCWrapperMock{}
+	mockClientWrapper := &pulsartestutils.CustomRPCWrapperMock{}
 
 	done := make(chan *rpc.Call, 1)
 	done <- &rpc.Call{Reply: &Payload{Body: HandshakePayload{}}}
@@ -234,16 +234,16 @@ func TestPulsar_EstablishConnection_IsNotInitialised_Success(t *testing.T) {
 	pulsar := &Pulsar{
 		Neighbours:       map[string]*Neighbour{},
 		PrivateKey:       mainPrivateKey,
-		EntropyGenerator: pulsartestutil.MockEntropyGenerator{},
+		EntropyGenerator: pulsartestutils.MockEntropyGenerator{},
 	}
 	firstPrivateKey, _ := ecdsahelper.GeneratePrivateKey()
 	expectedNeighbourKey, _ := ecdsahelper.ExportPublicKey(&firstPrivateKey.PublicKey)
-	payload := Payload{Body: HandshakePayload{Entropy: pulsartestutil.MockEntropy}}
+	payload := Payload{Body: HandshakePayload{Entropy: pulsartestutils.MockEntropy}}
 	sign, err := signData(firstPrivateKey, payload.Body)
 	payload.Signature = sign
 	payload.PublicKey = expectedNeighbourKey
 
-	mockClientWrapper := &pulsartestutil.CustomRPCWrapperMock{}
+	mockClientWrapper := &pulsartestutils.CustomRPCWrapperMock{}
 
 	done := make(chan *rpc.Call, 1)
 	done <- &rpc.Call{Reply: &payload}
@@ -263,7 +263,7 @@ func TestPulsar_CheckConnectionsToPulsars_NoProblems(t *testing.T) {
 	done <- &rpc.Call{}
 	replyChan := &rpc.Call{Done: done}
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("IsInitialised").Return(true)
 	mockClientWrapper.On("Go", HealthCheck.String(), nil, nil, mock.Anything).Return(replyChan)
 
@@ -282,7 +282,7 @@ func TestPulsar_CheckConnectionsToPulsars_NoProblems(t *testing.T) {
 }
 
 func TestPulsar_CheckConnectionsToPulsars_NilClient_FirstConnectionFailed(t *testing.T) {
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("IsInitialised").Return(false)
 	mockClientWrapper.On("Lock")
 	mockClientWrapper.On("Unlock")
@@ -305,7 +305,7 @@ func TestPulsar_CheckConnectionsToPulsars_NilClient_SecondConnectionFailed(t *te
 	done := make(chan *rpc.Call, 1)
 	done <- &rpc.Call{Error: errors.New("test error")}
 	replyChan := &rpc.Call{Done: done}
-	mockClientWrapper := &pulsartestutil.CustomRPCWrapperMock{}
+	mockClientWrapper := &pulsartestutils.CustomRPCWrapperMock{}
 	mockClientWrapper.Done = replyChan
 	isInitTimeCalled := 0
 	mockClientWrapper.IsInitFunc = func() bool {
@@ -374,7 +374,7 @@ func TestPulsar_StartConsensusProcess_Success(t *testing.T) {
 	mockSwitcher := MockStateSwitcher{}
 	mockSwitcher.On("SwitchToState", WaitingForEntropySigns, nil)
 	pulsar := &Pulsar{
-		EntropyGenerator: pulsartestutil.MockEntropyGenerator{},
+		EntropyGenerator: pulsartestutils.MockEntropyGenerator{},
 		PrivateKey:       mainPrivateKey,
 		OwnedBftRow:      map[string]*BftCell{},
 	}
@@ -392,7 +392,7 @@ func TestPulsar_StartConsensusProcess_Success(t *testing.T) {
 }
 
 func TestPulsar_broadcastSignatureOfEntropy_StateFailed(t *testing.T) {
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("Go", mock.Anything, mock.Anything, mock.Anything, mock.Anything)
 
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, StateSwitcher: &MockStateSwitcher{}}
@@ -414,11 +414,11 @@ func TestPulsar_broadcastSignatureOfEntropy_SendToNeighbours(t *testing.T) {
 	mainPrivateKey, err := ecdsahelper.GeneratePrivateKey()
 	assert.NoError(t, err)
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, PrivateKey: mainPrivateKey, StateSwitcher: &MockStateSwitcher{}}
 	pulsar.StateSwitcher.setState(WaitingForStart)
 	pulsar.ProcessingPulseNumber = 123
-	pulsar.GeneratedEntropySign = pulsartestutil.MockEntropy[:]
+	pulsar.GeneratedEntropySign = pulsartestutils.MockEntropy[:]
 	pulsar.Neighbours["1"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "first"}
 	pulsar.Neighbours["2"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "second"}
 
@@ -433,7 +433,7 @@ func TestPulsar_broadcastSignatureOfEntropy_SendToNeighbours(t *testing.T) {
 }
 
 func TestPulsar_broadcastVector_StateFailed(t *testing.T) {
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("Go", mock.Anything, mock.Anything, mock.Anything, mock.Anything)
 
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, OwnedBftRow: map[string]*BftCell{}, StateSwitcher: &MockStateSwitcher{}}
@@ -455,11 +455,11 @@ func TestPulsar_broadcastVector_SendToNeighbours(t *testing.T) {
 	mainPrivateKey, err := ecdsahelper.GeneratePrivateKey()
 	assert.NoError(t, err)
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, PrivateKey: mainPrivateKey, OwnedBftRow: map[string]*BftCell{}, StateSwitcher: &MockStateSwitcher{}}
 	pulsar.StateSwitcher.setState(WaitingForStart)
-	pulsar.GeneratedEntropySign = pulsartestutil.MockEntropy[:]
-	pulsar.GeneratedEntropy = pulsartestutil.MockEntropy
+	pulsar.GeneratedEntropySign = pulsartestutils.MockEntropy[:]
+	pulsar.GeneratedEntropy = pulsartestutils.MockEntropy
 	pulsar.Neighbours["1"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "first"}
 	pulsar.Neighbours["2"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "second"}
 
@@ -474,7 +474,7 @@ func TestPulsar_broadcastVector_SendToNeighbours(t *testing.T) {
 }
 
 func TestPulsar_broadcastEntropy_StateFailed(t *testing.T) {
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	mockClientWrapper.On("Go", mock.Anything, mock.Anything, mock.Anything, mock.Anything)
 
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, OwnedBftRow: map[string]*BftCell{}, StateSwitcher: &MockStateSwitcher{}}
@@ -496,11 +496,11 @@ func TestPulsar_broadcastEntropy_SendToNeighbours(t *testing.T) {
 	mainPrivateKey, err := ecdsahelper.GeneratePrivateKey()
 	assert.NoError(t, err)
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, PrivateKey: mainPrivateKey, StateSwitcher: &MockStateSwitcher{}}
 	pulsar.StateSwitcher.setState(WaitingForStart)
 	pulsar.ProcessingPulseNumber = 123
-	pulsar.GeneratedEntropy = pulsartestutil.MockEntropy
+	pulsar.GeneratedEntropy = pulsartestutils.MockEntropy
 	pulsar.Neighbours["1"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "first"}
 	pulsar.Neighbours["2"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "second"}
 
@@ -547,7 +547,7 @@ func TestPulsar_sendVector_TwoPulsars(t *testing.T) {
 	mainPrivateKey, err := ecdsahelper.GeneratePrivateKey()
 	assert.NoError(t, err)
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	pulsar := Pulsar{
 		Neighbours:  map[string]*Neighbour{},
 		PrivateKey:  mainPrivateKey,
@@ -555,7 +555,7 @@ func TestPulsar_sendVector_TwoPulsars(t *testing.T) {
 		bftGrid:     map[string]map[string]*BftCell{},
 	}
 
-	pulsar.GeneratedEntropy = pulsartestutil.MockEntropy
+	pulsar.GeneratedEntropy = pulsartestutils.MockEntropy
 	pulsar.Neighbours["1"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "first"}
 	mockClientWrapper.On("Go", ReceiveVector.String(), mock.Anything, nil, (chan *rpc.Call)(nil)).Return(replyChan)
 
@@ -595,9 +595,9 @@ func TestPulsar_sendEntropy_TwoPulsars(t *testing.T) {
 	mainPrivateKey, err := ecdsahelper.GeneratePrivateKey()
 	assert.NoError(t, err)
 
-	mockClientWrapper := &pulsartestutil.MockRPCClientWrapper{}
+	mockClientWrapper := &pulsartestutils.MockRPCClientWrapper{}
 	pulsar := Pulsar{Neighbours: map[string]*Neighbour{}, PrivateKey: mainPrivateKey, OwnedBftRow: map[string]*BftCell{}}
-	pulsar.GeneratedEntropy = pulsartestutil.MockEntropy
+	pulsar.GeneratedEntropy = pulsartestutils.MockEntropy
 	pulsar.Neighbours["1"] = &Neighbour{OutgoingClient: mockClientWrapper, ConnectionAddress: "first"}
 	mockClientWrapper.On("Go", ReceiveEntropy.String(), mock.Anything, nil, (chan *rpc.Call)(nil)).Return(replyChan)
 
@@ -631,7 +631,7 @@ func TestPulsar_verify_Standalone_Success(t *testing.T) {
 	pulsar := &Pulsar{StateSwitcher: mockSwitcher}
 	pulsar.StateSwitcher.setState(Verifying)
 	pulsar.PublicKeyRaw = "testKey"
-	pulsar.GeneratedEntropy = pulsartestutil.MockEntropy
+	pulsar.GeneratedEntropy = pulsartestutils.MockEntropy
 	pulsar.OwnedBftRow = map[string]*BftCell{}
 	pulsar.bftGrid = map[string]map[string]*BftCell{}
 
@@ -639,7 +639,7 @@ func TestPulsar_verify_Standalone_Success(t *testing.T) {
 
 	mockSwitcher.AssertCalled(t, "SwitchToState", SendingPulse, nil)
 	assert.Equal(t, "testKey", pulsar.PublicKeyRaw)
-	assert.Equal(t, core.Entropy(pulsartestutil.MockEntropy), pulsar.GeneratedEntropy)
+	assert.Equal(t, core.Entropy(pulsartestutils.MockEntropy), pulsar.GeneratedEntropy)
 }
 
 func TestPulsar_verify_NotEnoughForConsensus_Success(t *testing.T) {
@@ -690,7 +690,7 @@ func TestPulsar_verify_Success(t *testing.T) {
 	privateKeySecond, _, publicKeySecond := generatePrivateAndConvertPublic(t)
 	privateKeyThird, _, publicKeyThird := generatePrivateAndConvertPublic(t)
 
-	clientMock := pulsartestutil.MockRPCClientWrapper{}
+	clientMock := pulsartestutils.MockRPCClientWrapper{}
 	clientMock.On("IsInitialised").Return(true)
 	pulsar := &Pulsar{
 		StateSwitcher:                  mockSwitcher,
