+++ pulsar/pulsartestutils/mockrpcclientwrapper.go
@@ -15,7 +15,7 @@
  */
 
 // Package pulsartestutil - test utils for pulsar package
-package pulsartestutil
+package pulsartestutils
 
 import (
 	"net/rpc"
